@@ -1,31 +1,45 @@
-[tool.poetry]
+[project]
 name = "digiarch"
-version = "5.2.5"
+version = "6.0.0"
 description = "Tools for the Digital Archive Project at Aarhus Stadsarkiv"
-authors = ["Aarhus Stadsarkiv <stadsarkiv@aarhus.dk>"]
-license = "GPL-3.0"
+authors = [{ name = "Aarhus Stadsarkiv", email = "stadsarkiv@aarhus.dk" }]
+requires-python = ">=3.12"
 readme = "README.md"
-homepage = "https://stadsarkiv.aarhus.dk/"
-repository = "https://github.com/aarhusstadsarkiv/digiarch"
-keywords = ["archiving", "preservation", "identification"]
-include = ["pyproject.toml"]
+license = "GPL-3.0"
+keywords = [
+    "archiving",
+    "preservation",
+    "identification",
+]
+dependencies = [
+    "acacore",
+    "patool>=3.1.0,<4",
+    "tnefparse>=1.4.0,<2",
+    "extract-msg>=0.52.0,<0.53",
+    "chardet>=5.2.0,<6",
+]
 
-[tool.poetry.dependencies]
-python = "^3.11"
-acacore = {git = "https://github.com/aarhusstadsarkiv/acacore.git", tag = "v4.1.2"}
-patool = "^3.1.0"
-tnefparse = "^1.4.0"
-extract-msg = "^0.52.0"
-chardet = "^5.2.0"
+[project.urls]
+Homepage = "https://stadsarkiv.aarhus.dk/"
+Repository = "https://github.com/aarhusstadsarkiv/digiarch"
 
-[tool.poetry.group.dev.dependencies]
-pytest = "^8.3.2"
-ruff = "^0.5.6"
-coverage = "^7.6.1"
-types-click = "^7.1.8"
-
-[tool.poetry.scripts]
+[project.scripts]
 digiarch = "digiarch.cli:app"
+
+[dependency-groups]
+dev = [
+    "pytest>=8.3.4",
+    "ruff>=0.9.3",
+    "coverage>=7.6.10",
+    "types-click>=7.1.8",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.uv.sources]
+acacore = { git = "https://github.com/aarhusstadsarkiv/acacore.git", tag = "v5.0.0" }
 
 [tool.ruff]
 line-length = 120
@@ -139,11 +153,6 @@ ignore-decorators = ["pytest.fixture"]
 [tool.ruff.lint.pycodestyle]
 max-doc-length = 300
 
-[build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
-
 [tool.pytest.ini_options]
 asyncio_mode="strict"
-pythonpath = ["."]
-
+pythonpath = ["src"]
