@@ -1,5 +1,4 @@
 from typing import ClassVar
-from typing import Type
 
 from acacore.utils.decorators import docstring_format
 from click import argument
@@ -41,7 +40,7 @@ class ShellChoice(CompleteChoice):
 @argument("shell", type=ShellChoice(), callback=lambda _c, _p, v: get_completion_class(v))
 @pass_context
 @docstring_format(shells="\n".join(f"* {s.value}\t{s.help}" for s in ShellChoice.completion_items))
-def cmd_completions(ctx: Context, shell: Type[ShellComplete] | None):
+def cmd_completions(ctx: Context, shell: type[ShellComplete] | None):
     """
     Generate completion scripts for your shell.
 
