@@ -1,5 +1,5 @@
+from collections.abc import Callable
 from typing import Any
-from typing import Callable
 
 from acacore.database import FilesDB
 from acacore.database.table import Table
